move getWarnedUsers into warnings.js
